replace all torch.bfloat16 with torch.float16 as a workaround for: RuntimeError: "ms_deform_attn_forward_cuda" not implemented for 'BFloat16'

+++ demo/backend/server/inference/predictor.py
@@ -93,7 +93,7 @@ class InferenceAPI:
 
     def autocast_context(self):
         if self.device.type == "cuda":
-            return torch.autocast("cuda", dtype=torch.bfloat16)
+            return torch.autocast("cuda", dtype=torch.float16)
         else:
             return contextlib.nullcontext()
