+++ grounded_sam2_tracking_demo_custom_video_input_gd1.0_local_model.py
@@ -112,7 +112,7 @@ OBJECTS = class_names
 print(OBJECTS)
 
 # FIXME: figure how does this influence the G-DINO model
-torch.autocast(device_type=DEVICE, dtype=torch.bfloat16).__enter__()
+torch.autocast(device_type=DEVICE, dtype=torch.float16).__enter__()
 
 if torch.cuda.get_device_properties(0).major >= 8:
     # turn on tfloat32 for Ampere GPUs (https://pytorch.org/docs/stable/notes/cuda.html#tensorfloat-32-tf32-on-ampere-devices)
