+++ grounded_sam2_tracking_demo_with_gd1.5.py
@@ -19,7 +19,7 @@ from utils.video_utils import create_video_from_images
 Step 1: Environment settings and model initialization for SAM 2
 """
 # use bfloat16 for the entire notebook
-torch.autocast(device_type="cuda", dtype=torch.bfloat16).__enter__()
+torch.autocast(device_type="cuda", dtype=torch.float16).__enter__()
 
 if torch.cuda.get_device_properties(0).major >= 8:
     # turn on tfloat32 for Ampere GPUs (https://pytorch.org/docs/stable/notes/cuda.html#tensorfloat-32-tf32-on-ampere-devices)
